+++ spec/12-encoding_spec.lua
@@ -1,3 +1,4 @@
+_G._TEST = true
 local encoding = require "openapi2kong.encoding"
 
 assert:set_parameter("TableFormatLevel", 5)
@@ -6,13 +7,13 @@ describe("[encoding]", function()
 
   it("requires a table parameter as spec", function()
     local ok, err = encoding("myProperty", "lalala", nil, {})
-    assert.equal("a encoding object expects a table as spec, but got string", err)
+    assert.equal("a encoding object expects a table as spec, but got string (origin: PARENT:encoding[myProperty])", err)
     assert.falsy(ok)
   end)
 
   it("requires a string parameter as property_name", function()
     local ok, err = encoding(123, {}, nil, {})
-    assert.equal("a encoding object expects a string as property_name, but got number", err)
+    assert.equal("a encoding object expects a string as property_name, but got number (origin: PARENT:encoding[123])", err)
     assert.falsy(ok)
   end)
 
@@ -32,7 +33,7 @@ describe("[encoding]", function()
     local result, err = encoding("myProperty", {
         headers = "lalala",
       }, nil, {})
-    assert.equal("a headers object expects a table, but got string", err)
+    assert.equal("a headers object expects a table, but got string (origin: PARENT:encoding[myProperty]:headers)", err)
     assert.is_nil(result)
   end)