+++ src/openapi2kong/encodings.lua
@@ -6,6 +6,11 @@ local TYPE_NAME = ({...})[1]:match("openapi2kong%.([^%.]+)$")  -- grab type-name
 local mt = require("openapi2kong.common").create_mt(TYPE_NAME)
 
 
+function mt:get_trace()
+  return ""
+end
+
+
 function mt:validate()
 
   if type(self.spec) ~= "table" then
@@ -24,7 +29,7 @@ function mt:post_validate()
 end
 
 
--- returns an array with path entries
+-- returns an array with encoding entries
 local function parse(spec, options, parent)
 
   local self = setmetatable({
@@ -35,7 +40,7 @@ local function parse(spec, options, parent)
 
   local ok, err = self:validate()
   if not ok then
-    return ok, err
+    return ok, self:log_message(err)
   end
 
   self.encodings = {}
@@ -49,7 +54,7 @@ local function parse(spec, options, parent)
 
   ok, err = self:post_validate()
   if not ok then
-    return ok, err
+    return ok, self:log_message(err)
   end
 
   return self