+++ src/openapi2kong/mediaType.lua
@@ -3,6 +3,11 @@ local TYPE_NAME = ({...})[1]:match("openapi2kong%.([^%.]+)$")  -- grab type-name
 local mt = require("openapi2kong.common").create_mt(TYPE_NAME)
 
 
+function mt:get_trace()
+  return self.mediatype
+end
+
+
 function mt:validate()
 
   if type(self.mediatype) ~= "string" then
@@ -36,7 +41,7 @@ local function parse(mediatype, spec, options, parent)
 
   local ok, err = self:validate()
   if not ok then
-    return ok, err
+    return ok, self:log_message(err)
   end
 
   if self.spec.schema then
@@ -57,7 +62,7 @@ local function parse(mediatype, spec, options, parent)
 
   ok, err = self:post_validate()
   if not ok then
-    return ok, err
+    return ok, self:log_message(err)
   end
 
   return self