+++ src/openapi2kong/openapi.lua
@@ -3,6 +3,11 @@ local TYPE_NAME = ({...})[1]:match("openapi2kong%.([^%.]+)$")  -- grab type-name
 local mt = require("openapi2kong.common").create_mt(TYPE_NAME)
 
 
+function mt:get_trace()
+  return (self.spec.info or {}).title or ""
+end
+
+
 function mt:validate()
 
   if type(self.spec) ~= "table" then
@@ -119,7 +124,7 @@ local function parse(spec, options, parent)
 
   local ok, err = self:validate()
   if not ok then
-    return ok, err
+    return ok, self:log_message(err)
   end
 
   for _, property in ipairs { "servers", "paths" } do
@@ -146,7 +151,7 @@ local function parse(spec, options, parent)
 
   ok, err = self:post_validate()
   if not ok then
-    return ok, err
+    return ok, self:log_message(err)
   end
 
   return self