+++ src/openapi2kong/path.lua
@@ -7,6 +7,11 @@ local TYPE_NAME = ({...})[1]:match("openapi2kong%.([^%.]+)$")  -- grab type-name
 local mt = require("openapi2kong.common").create_mt(TYPE_NAME)
 
 
+function mt:get_trace()
+  return self.path
+end
+
+
 function mt:validate()
 
   if type(self.path) ~= "string" then
@@ -43,7 +48,7 @@ local function parse(path, spec, options, parent)
   do
     local ok, err = self:dereference()
     if not ok then
-      return ok, err
+      return ok, self:log_message(err)
     end
     -- prevent accidental access to non-dereferenced spec table
     spec = nil -- luacheck: ignore
@@ -51,7 +56,7 @@ local function parse(path, spec, options, parent)
 
   local ok, err = self:validate()
   if not ok then
-    return ok, err
+    return ok, self:log_message(err)
   end
 
   for _, property in ipairs { "servers", "parameters" } do
@@ -91,7 +96,7 @@ local function parse(path, spec, options, parent)
 
   ok, err = self:post_validate()
   if not ok then
-    return ok, err
+    return ok, self:log_message(err)
   end
 
   return self