+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 0.x.x Unreleased
+
+ - added a trace to errors to identify to originating yaml/json element that
+   caused the error.
+
 ## 0.1.0
 
 - initial version